Added two scripts to perform some regression testing:
take_model_snapshot.py
Takes a snapshot of a how Delphi works at a particular point of its implementation
A snapshot taken before and a snapshot taken after a code change such as an
optimization or refactoring could be used to verify whether the accuracy of the
code got changed as a byproduct of the changes made to the code. If there is no
change at all, the before and after snaps should be identical.

take_model_snapshot.py
Compares two snapshots of how Delphi worked at two points of its implementation
A snapshot taken before and a snapshot taken after a code change such as an
optimization or refactoring could be compared to verify whether the accuracy of the
code got changed as a byproduct of the changes made to the code. If there is no
change at all, the before and after snaps should be identical and the plots produced
by this script should be a single bar at zero.