Resolving multiple uses of the "res" variable.
The member variable "res" was used for the sampling resolution and the
number of KDE kernels. Therefore, Delphi ended up using the same value
for both.

Separated these usages by introducing a new member variable
"n_kde_kernels". The resolution was done in the least code change way
and hence not the most elegant approach. Ideally, n_kde_kernels should
be an argument for the AnalysisGraph::run_train_model() method.