Delphi eliminate logpdf recalculation (#508)
* Eliminated calculating logpdf for an edge twice.

We need the before and after logpdf values for an edge that gets
perturbed during the sampling. Previously, we were calculating both of
them each time. Now the program memorizes the previous value eliminating
the need to compute it twice.

* Profiling improvements:

1. Measuring CPU time.
2. Micro-timing the MCMC loop.
3. Added a cmake compile flag to turn micro-timing code on and off.
4. Changed the clock used for duration measurements from
high_resolution_clock to steady_clock.
5. Changed timing measurement unit into nanoseconds from milliseconds.
6. Updated the timing analysis plotting script and added a 3D plotting
script. But the 3D plots were hard to interpret.

* Added a separate profiling section to AnalysisGraph

Also added a script to assemble and plot timing results.

* Fixed a bug due to a typo

* Updated the Delphi profiler to profile matrix exponential and updating
the transition matrix after perturbing a theta.

Co-authored-by: Manujinda Wathugala <manujinda@email.arizona.edu>