Fixed some bugs and a slowness issue
1. Bugs
When deserializing a trained AnalysisGraph object, the sampled edge
weights (angles) were not copied into the new object.

When returning the trained model status, the sampled edge weights were
reported as slopes. Changed that to angles.

C++ REST API: When creating a model, the number of KDE Kernels was not
set. So, the default value (1000) was used always.

2. Slowness
Added some timing code to model status endpoint of the C++ REST API.

Fixed the slowness issue of the AnalysisGraph deserialization.