Enriched the copy constructor
Previously, the copy constructor was written to copy an untrained
AnalysisGraph. Therefore, it did not copy any members than would get
populated while training. After writing the copy assignment operator, to
better align with it and to prevent the danger of an inconsistent state
that could arise if someone copies a trained model, completed the copy
constructor to copy all the members that are copyable.