Fixed a bug related to freezing edges.
1. When we
create a model
train it
serialize it
deserialize it
freeze an edge
check model status,
the frozen edge did not gave the value set for it. The sampled thetas
for a edge was not cleared when we train a model for again and again and
hence the values sampled before freezing too were included when
calculating the average edge weight.

2. Even after fixing the first bug, frozen edge were giving different
weights each time the model was trained. When checked, the REST API did
not properly froze the edge. In the REST API the deserialized
AnalysisGraph object was passed by value to another function that called
the freeze_edge_weight() method on it. Since it was called on a copy,
the changes were not applied to the object that got trained and written
to the db.