Enriched model status reporting.
Delphi was reporting that a model is ready when model status REST API
endpoint is queried while the model is being trained after a edit edges
call.

This happened since we did not write the edge weight set model back to
delphi.db and we do not reset the model status as not trained just after
an edge was edited.

Fixed this issue.

However, could not test with the REST server as it crashes when trying
to run even without any of the updates.