Freezing and unfreezing edges
1. Allowed re-freezing an already frozen edge to a new value. Since we
are warm starting training, the newly sampled edge weights for
non-frozen edges could end up being close to the previously trained
model.

2. Overloaded the AnalysisGraph::freeze_edge_weight() method to unfreeze
an edge by providing a polarity of 0. In this case warm starting the
training based on the MAP estimate of the previous training run is
disabled to enable fresh training.