Removed the necessity to pass a dummy zero matrix into AnalysisGraph::assemble_head_node_modeling_LDS() method to represent the transition matrix that models the relationships between concepts.
Earlier I was thinking of reusing this method to assemble the final
complete LDS with all the data.

Now, this method only assembles the seasonal head node modeling
portion of the complete LDS. It leaves zeros where the relationships
between concepts should be at. We have to insert the transition matrix
that models these relationships into the correct place.