Making matrix exponential more efficient
When we use the Fourier decomposition based head node model, the head
node related transition matrix portion does not change after fitting.
Also, it is independent of the rest of the model. Therefore, we do not
need to take the matrix exponential of this portion again and again.

So, for all the unique gaps between observation time points, we compute
the matrix exponential of the Fourier decomposition based base head node
model transition matrix and remember them. When we are advancing the
system to compute the log likelihood, we merge these with the base
transition matrix that describe the relationships between concepts
exponentiated with the respective gap to get the full transition matrix.

When projecting, we combine the two base matrices and then take the
matrix exponential.