Fixed a bug.
When a model is trained more than once, the program was crashing.
This was due to the state of the head node not being cleaned before
fitting the Fourier decomposition based seasonal head node model. This
caused partitions to accumulate data points multiple times.

Implemented a method to clear the state of a node.

Also initialized log_likelihood to 0 before each training run.