Fixed a division by zero problem
When there are no midpoints calculated for a seasonal head node, we
cannot compute the root mean squared error to choose the best number of
sinusoidals to be used to fit it via Fourier decomposition. We get a
zero length error vector and we were dividing by that size.

Fixed it.