Fixed a bug in MCMC sampling.
When we sample a theta, we update the transition matrix and then compute
the matrix exponentials for different gaps before evaluating the log
likelihood. If that sample gets rejected, we revert the transition
matrix but do not re-compute the matrix exponentials.

If the next sample happens to be a theta, this cause no issue as we
again update the transition matrix and compete the matrix exponentials
for the updated transition matrix.

However, when the next sample is a derivative, we do not update the
transition matrix and do not compute the matrix exponentials. Therefore,
we end up using the matrix exponentials of the previous rejected theta
sample to evaluate the log likelihood of the theta sample. This is as if
we accepted the previous rejected theta sample temporarily. If we end up
accepting the sampled derivative, we end up accepting the previously
rejected theta sample along with it.

Fixed this bug.

Along with this fix, implemented a way to reduce some heavy computation.
When a theta sample is rejected, we have to revert the transition
matrix. This includes the costly chain rule evaluation. To avoid this
computation, made the program remember the previous transition matrix
before updating it according to the sampled theta.

When a transition matrix is reverted, we have to re-compute its matrix
exponentials for different gaps, which is another costly computation.
Avoided this computation by making the program remember the previously
computed matrix exponentials before the matrix exponentials are computed
for the updated matrix.